+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 * Users listing
 * Collections listing
 
+### Change
+
+* search documents returns empty `Vec` instead of None in case no documents have been found.
+
 
 ## [0.3.7] - 2019-06-14